build(native): exclude daemon-only sources from `libsentry.so`
`sentry-crash` previously copied all sources from the `sentry` target,
so daemon-only crash processing, minidump, and remote unwinding sources
were also built into `libsentry.so`.

Add those sources directly to `sentry-crash` instead. Keep `daemon_start`
in the native backend because launching the handler remains a client-side
responsibility.

This removes the following Linux x86_64 stack frames from `libsentry.so`:

- `sentry__write_minidump`: 1,216,336 bytes
- `sentry__process_crash`: 18,368 bytes
- `enrich_frame_with_symbol`: 14,592 bytes
- `write_attachment_to_envelope`: 8,448 bytes

The source split does not preclude a potential future in-process mode.
