+++ src/backends/native/sentry_crash_daemon.h
@@ -14,29 +14,6 @@
 struct sentry_options_s;
 
 /**
- * Start crash daemon for monitoring app process
- * This forks a child process (Unix) or creates a new process (Windows) that
- * waits for crashes
- *
- * @param app_pid Parent application process ID
- * @param app_tid Parent application thread ID
- * @param notify_handle Crash notification handle
- * @param ready_handle Ready signal handle
- * @return Daemon PID on success, -1 on failure
- */
-#if defined(SENTRY_PLATFORM_LINUX) || defined(SENTRY_PLATFORM_ANDROID)
-pid_t sentry__crash_daemon_start(pid_t app_pid, uint64_t app_tid,
-    int notify_eventfd, int ready_eventfd, const char *handler_path);
-#elif defined(SENTRY_PLATFORM_MACOS)
-pid_t sentry__crash_daemon_start(pid_t app_pid, uint64_t app_tid,
-    int notify_pipe_read, int ready_pipe_write, int shm_fd,
-    const char *handler_path);
-#elif defined(SENTRY_PLATFORM_WINDOWS)
-pid_t sentry__crash_daemon_start(pid_t app_pid, uint64_t app_tid,
-    HANDLE event_handle, HANDLE ready_event_handle, const char *handler_path);
-#endif
-
-/**
  * Daemon main loop (runs in forked child on Unix, or separate process on
  * Windows)
  * @param app_pid Parent process ID
